// infrastructure/frameworks/frontend/next/src/types/image-asset.types.ts
export type ImageAssetHttpItem = {
  id: string;
  title: string;
  originalFileName: string;
  sourceMimeType: string;
  sourceSizeBytes: number;
  webpSizeBytes: number;
  width: number;
  height: number;
  webpFileName: string;
  webpPath: string;
  webpUrl: string;
  createdAt: string;
  updatedAt: string;
};